refactor: make a backend two files of its own instead of a section of four (#15)
Adding a VPN tool was documented as "one file and one line". It was five files
in nine places: a section appended to a 1366-line Model.js, cases added to a
768-line tests/run.js, two lists and an if-chain in VpnController.qml, an enum
and three prose lines in manifest.json, and a sentence in Panel.qml. Most of
that is shared-file editing, which is the kind a contributor cannot do
confidently on a first change and the kind that conflicts when two people add
tools at once.

Model.js is now model/Shared.js plus one file per tool, sliced on the banner
comments that were already there. No logic moved except favoriteCodes, which
lived under the Proton banner while Mullvad and Windscribe both called it, and
is shared by any reading. Each backend imports only Shared and its own model, so
nothing can quietly grow a dependency on another tool's parser. The QML engine
resolves `.import "Shared.js" as Shared` from a `.pragma library`, which was
worth checking before relying on it.

tests/run.js is now a harness plus one file per tool, picked up by glob. The
suite ran unchanged against the split files before the tests themselves were
moved, so the 62 existing assertions are what says the move was faithful.

Two things the panel had hardcoded now come from the backends. preferredBackend
mapped labels to ids with a four-line if-chain; it reads the backends instead,
which is where a label is already declared. The "install something" line named
its tools in prose and had gone a whole release without mentioning Windscribe —
exactly the failure a derived list cannot have. Backends declare `installNames`
for it, because the label is not always the answer: nobody installs
NetworkManager to get a VPN, so that one offers OpenVPN and WireGuard.

What remains manual is the preferredBackend enum in manifest.json. Omarchy's
settings dialog reads that JSON before any QML runs, so nothing at runtime can
fill it in. It is one line, and now it is written down as the one line.

CONTRIBUTING.md covers the parts that are not derivable from the code: that dev
is the branch to open against, that commit messages drive the changelog and the
version, that model/ does not hot-reload, and that a stale parser wants the real
CLI output pasted into a test rather than a tidied version of it. ARCHITECTURE.md
stops claiming one file and one line.

Closes #12.
