--- Model.js
@@ -1,1366 +1,0 @@
-.pragma library
-
-// Nerd Font glyphs are built from codepoints instead of raw characters so the
-// file survives editing tools that mangle multi-byte sequences.
-var GLYPH_VPN = String.fromCodePoint(0xF0582)
-var GLYPH_CHECK = String.fromCodePoint(0xF012C)
-var GLYPH_LOCK = String.fromCodePoint(0xF033E)
-var GLYPH_BOLT = String.fromCodePoint(0xF04C5)
-var GLYPH_DICE = String.fromCodePoint(0xF01D5)
-var GLYPH_SWAP = String.fromCodePoint(0xF04E1)
-var GLYPH_SHIELD = String.fromCodePoint(0xF0498)
-var GLYPH_CHEVRON_DOWN = String.fromCodePoint(0xF0140)
-var GLYPH_CHEVRON_UP = String.fromCodePoint(0xF0143)
-var GLYPH_COG = String.fromCodePoint(0xF0493)
-
-// ----------------------------------------------------------------- shared
-
-// The exit address lookup answers with a bare address and nothing else. A
-// captive portal's login page, a proxy's error body, or anything else that
-// came back with it is not an answer — and this is the one number a user reads
-// to decide whether the tunnel is carrying their traffic, so rendering
-// whatever arrived would be the widget confirming a route it never saw.
-// Returns "" for anything that is not an address literal.
-function parsePublicIp(raw) {
-  var text = String(raw || "").trim()
-  // Longest possible IPv6 text form; anything longer is not an address.
-  if (text === "" || text.length > 45) return ""
-
-  if (/^\d{1,3}(\.\d{1,3}){3}$/.test(text)) {
-    var octets = text.split(".")
-    for (var i = 0; i < octets.length; i++) {
-      if (parseInt(octets[i], 10) > 255) return ""
-    }
-    return text
-  }
-
-  // IPv6 has too many legal spellings to re-implement here, so this checks the
-  // alphabet and the shape rather than the grouping.
-  if (/^[0-9a-fA-F:]+$/.test(text) && text.indexOf("::") === text.lastIndexOf("::")) {
-    if (text.indexOf(":") !== -1 && !/:::/.test(text)) return text.toLowerCase()
-  }
-  return ""
-}
-
-function elide(text, limit) {
-  var value = String(text || "").replace(/\s+/g, " ").trim()
-  return value.length > limit ? value.substring(0, limit - 1) + "…" : value
-}
-
-function detail(label, value) {
-  return { label: label, value: String(value || "") }
-}
-
-// A tool setting the panel can flip. The value is always what the tool last
-// reported, never something the widget stores — nothing here owns a setting.
-function toggle(key, label, description, value) {
-  return { key: key, label: label, detail: description, value: value === true, busy: false }
-}
-
-// A toggle the user just flipped shows the new position and a busy marker until
-// the tool confirms it, so the switch does not sit still under the click.
-function applyPendingToggles(toggles, pending) {
-  if (!pending) return toggles
-  return toggles.map(function(entry) {
-    if (pending[entry.key] === undefined) return entry
-    return { key: entry.key, label: entry.label, detail: entry.detail, value: pending[entry.key] === true, busy: true }
-  })
-}
-
-// --------------------------------------------------------- widget settings
-
-// Which tools the user told the widget to ignore, stored as one comma-separated
-// string so the setting stays hand-editable in shell.json and in Omarchy's own
-// settings dialog, which has no array field.
-function parseBackendIds(raw) {
-  var ids = []
-  var parts = String(raw || "").split(",")
-  for (var i = 0; i < parts.length; i++) {
-    var id = parts[i].trim().toLowerCase()
-    if (id !== "" && ids.indexOf(id) === -1) ids.push(id)
-  }
-  return ids
-}
-
-function joinBackendIds(ids) {
-  return ids.join(",")
-}
-
-function toggleBackendId(ids, id) {
-  var next = []
-  var found = false
-  for (var i = 0; i < ids.length; i++) {
-    if (ids[i] === id) found = true
-    else next.push(ids[i])
-  }
-  if (!found) next.push(id)
-  return next
-}
-
-// ------------------------------------------------------------ Proton VPN
-
-// `protonvpn status` prints a plain-text block, not JSON:
-//
-//   Status: Connected
-//   Server: NL#42
-//   Country: Netherlands
-//   Load: 40%
-//   Protocol: WireGuard
-//
-// Labels vary between releases, so match on the leading key rather than on
-// line position, and treat anything unrecognized as noise.
-function parseProtonStatus(raw) {
-  var result = {
-    connected: false,
-    server: "",
-    country: "",
-    city: "",
-    load: "",
-    protocol: "",
-    ip: "",
-    statusText: ""
-  }
-
-  var lines = String(raw || "").split("\n")
-  for (var i = 0; i < lines.length; i++) {
-    var line = lines[i].trim()
-    if (line === "") continue
-
-    var separator = line.indexOf(":")
-    if (separator < 0) {
-      // Older builds print a bare "Connected to NL#42" sentence.
-      var loose = line.match(/^connected\s+to\s+(.+)$/i)
-      if (loose) {
-        result.connected = true
-        result.server = loose[1].trim()
-      }
-      continue
-    }
-
-    var key = line.substring(0, separator).trim().toLowerCase()
-    var value = line.substring(separator + 1).trim()
-    if (value === "") continue
-
-    if (key === "status") {
-      result.statusText = value
-      result.connected = /^connected/i.test(value)
-    } else if (key === "server" || key === "server name") {
-      result.server = value
-    } else if (key === "country") {
-      result.country = value
-    } else if (key === "city") {
-      result.city = value
-    } else if (key === "load" || key === "server load") {
-      result.load = value
-    } else if (key === "protocol") {
-      result.protocol = value
-    } else if (key === "ip" || key === "ip address") {
-      result.ip = value
-    }
-  }
-
-  if (result.statusText === "") result.statusText = result.connected ? "Connected" : "Disconnected"
-
-  // A connected CLI reports one combined field — "CH#1129 in Zurich,
-  // Switzerland" — rather than the separate Country/City keys older builds
-  // used. Split it so the panel can label the parts.
-  var located = result.server.match(/^(\S+)\s+in\s+(.+)$/i)
-  if (located) {
-    result.server = located[1]
-    var place = located[2].split(",")
-    if (result.city === "" && place.length > 1) result.city = place[0].trim()
-    if (result.country === "") result.country = (place.length > 1 ? place.slice(1).join(",") : place[0]).trim()
-  }
-
-  return result
-}
-
-// `protonvpn countries list` prints a two-column table padded with spaces,
-// preceded by an optional "Server list is outdated, updating..." notice and a
-// dashed rule under the header.
-function parseProtonCountries(raw) {
-  var countries = []
-  var lines = String(raw || "").split("\n")
-  for (var i = 0; i < lines.length; i++) {
-    var line = lines[i].trim()
-    if (line === "" || line.indexOf("---") === 0) continue
-
-    var match = line.match(/^(.+?)\s{2,}([A-Za-z]{2})$/)
-    if (!match) continue
-
-    var name = match[1].trim()
-    var code = match[2].toUpperCase()
-    if (name.toLowerCase() === "country") continue
-
-    countries.push({ name: name, code: code })
-  }
-  return countries
-}
-
-// `protonvpn config list` prints a padded two-column table under a dashed rule:
-//
-//   Setting                  Value
-//   -----------------------  ------------
-//   netshield                malware-only
-//   kill-switch              off
-function parseProtonConfig(raw) {
-  var values = {}
-  var loaded = false
-  var lines = String(raw || "").split("\n")
-
-  for (var i = 0; i < lines.length; i++) {
-    var line = lines[i].trim()
-    if (line === "" || line.indexOf("---") === 0) continue
-
-    var match = line.match(/^(\S+)\s{2,}(\S+)$/)
-    if (!match) continue
-
-    var key = match[1].toLowerCase()
-    if (key === "setting") continue
-
-    values[key] = match[2]
-    loaded = true
-  }
-
-  return { values: values, loaded: loaded }
-}
-
-// Three of Proton's settings read as plain switches. The rest — custom DNS,
-// NAT type, IPv6 — are not on/off questions and stay with the CLI.
-function protonToggles(config) {
-  if (!config || !config.loaded) return []
-
-  var netshield = String(config.values["netshield"] || "off")
-  var killSwitch = String(config.values["kill-switch"] || "off")
-
-  return [
-    toggle("kill-switch", "Kill switch",
-      killSwitch !== "off" && killSwitch !== "standard"
-        ? "Mode: " + killSwitch
-        : "Cut traffic if the tunnel drops",
-      killSwitch !== "off"),
-    toggle("netshield", "NetShield",
-      netshield !== "off" ? "Blocking: " + netshield : "Block ads, trackers, malware",
-      netshield !== "off"),
-    toggle("port-forwarding", "Port forwarding", "Open an inbound port for P2P",
-      String(config.values["port-forwarding"] || "off") === "on")
-  ]
-}
-
-// Turning one on means picking a value, since only "off" is shared. Switching
-// one off and on again should land back on the mode it had — "malware-only" is
-// a deliberate choice, and a switch that silently upgraded it to
-// "malware-ads-trackers" would be changing a setting nobody asked it to change.
-// `previousMode` is the last non-off value the tool reported.
-function protonToggleArgs(key, value, previousMode) {
-  var mode = String(previousMode || "")
-
-  if (key === "kill-switch") {
-    return ["config", "set", "kill-switch", value ? (mode !== "" ? mode : "standard") : "off"]
-  }
-  if (key === "netshield") {
-    return ["config", "set", "netshield", value ? (mode !== "" ? mode : "malware-ads-trackers") : "off"]
-  }
-  if (key === "port-forwarding") return ["config", "set", "port-forwarding", value ? "on" : "off"]
-  return []
-}
-
-// The modes worth remembering across an off/on round trip.
-function protonModes(config, known) {
-  var modes = {}
-  for (var key in known) modes[key] = known[key]
-  if (!config || !config.loaded) return modes
-
-  var keys = ["kill-switch", "netshield"]
-  for (var i = 0; i < keys.length; i++) {
-    var value = String(config.values[keys[i]] || "off")
-    if (value !== "off") modes[keys[i]] = value
-  }
-  return modes
-}
-
-function protonLocation(status) {
-  var parts = []
-  if (status.city !== "") parts.push(status.city)
-  if (status.country !== "") parts.push(status.country)
-  return parts.join(", ")
-}
-
-function protonSummary(status) {
-  if (!status.connected) return "Not connected"
-
-  var location = protonLocation(status)
-  if (location !== "" && status.server !== "") return status.server + " · " + location
-  if (location !== "") return location
-  if (status.server !== "") return status.server
-  return "Connected"
-}
-
-function protonDetails(status) {
-  if (!status.connected) return []
-
-  var rows = [detail("Server", status.server), detail("Location", protonLocation(status))]
-  if (status.load !== "") rows.push(detail("Load", status.load))
-  if (status.protocol !== "") rows.push(detail("Protocol", status.protocol))
-  return rows.filter(function(row) { return row.value !== "" })
-}
-
-function favoriteCodes(raw) {
-  var codes = []
-  var parts = String(raw || "").split(",")
-  for (var i = 0; i < parts.length; i++) {
-    var code = parts[i].trim().toUpperCase()
-    if (code !== "" && codes.indexOf(code) < 0) codes.push(code)
-  }
-  return codes
-}
-
-// Quick-connect rows above the country list. `args` is passed straight to
-// `protonvpn connect`.
-function protonQuickTargets() {
-  return [
-    { key: "fastest", label: "Fastest server", detail: "Best available worldwide", glyph: GLYPH_BOLT, args: [] },
-    { key: "p2p", label: "P2P server", detail: "Fastest server that allows P2P", glyph: GLYPH_SWAP, args: ["--p2p"] },
-    { key: "random", label: "Random server", detail: "Pick any available server", glyph: GLYPH_DICE, args: ["--random"] },
-    { key: "securecore", label: "Secure Core", detail: "Route through a hardened entry", glyph: GLYPH_LOCK, args: ["--securecore"] }
-  ]
-}
-
-// Favorites first, then the rest, with an optional substring filter over both
-// the country name and its code.
-function protonCountryTargets(countries, favorites, filter) {
-  var needle = String(filter || "").trim().toLowerCase()
-  var favored = []
-  var rest = []
-
-  for (var i = 0; i < countries.length; i++) {
-    var country = countries[i]
-    if (needle !== "") {
-      var haystack = (country.name + " " + country.code).toLowerCase()
-      if (haystack.indexOf(needle) < 0) continue
-    }
-
-    var target = {
-      key: "country:" + country.code,
-      label: country.name,
-      detail: country.code,
-      glyph: GLYPH_VPN,
-      args: ["--country", country.code]
-    }
-    if (needle === "" && favorites.indexOf(country.code) >= 0) favored.push(target)
-    else rest.push(target)
-  }
-
-  return favored.concat(rest)
-}
-
-// --------------------------------------------------------- NetworkManager
-//
-// OpenVPN and WireGuard both live here, because on a desktop both are
-// NetworkManager profiles: same listing call, same `connection up` verb, same
-// teardown. What differs is how NetworkManager types them — OpenVPN is a `vpn`
-// connection with a service-type plugin behind it, WireGuard is its own
-// connection type with the keys in the profile.
-
-// `nmcli -t` escapes literal colons as "\:", so split on the first unescaped
-// one rather than on every colon.
-function splitNmcliLine(line) {
-  var text = String(line || "")
-  for (var i = 0; i < text.length; i++) {
-    if (text[i] === "\\") { i++; continue }
-    if (text[i] === ":") return [unescapeNmcli(text.substring(0, i)), unescapeNmcli(text.substring(i + 1))]
-  }
-  return [unescapeNmcli(text), ""]
-}
-
-function unescapeNmcli(value) {
-  return String(value || "").replace(/\\(.)/g, "$1")
-}
-
-// Generated on the fly for a device someone else brought up, rather than a
-// profile on disk. Empty means the field was never asked for: nmcli refuses a
-// listing that names a field it does not know, so an nmcli too old for
-// FILENAME is retried without it — see NetworkManagerBackend. Such a
-// connection is kept, since a stray row beats a backend that lists nothing.
-function isVolatileConnection(filename) {
-  return String(filename || "").indexOf("/run/") === 0
-}
-
-// `nmcli -t -f NAME,UUID,TYPE,ACTIVE,FILENAME connection show` — one connection
-// per line. Two types are tunnels: `vpn` (an OpenVPN profile, or another
-// plugin's, which the second pass sorts out) and `wireguard`. Ethernet, wifi,
-// bridges and the rest are somebody else's business.
-//
-// FILENAME is what keeps other tools' tunnels out. When something brings up a
-// WireGuard interface itself — Mullvad's `wg0-mullvad`, say — NetworkManager
-// adopts the device and generates a volatile connection for it under
-// `/run/NetworkManager/`. Listing that would put the same tunnel on two chips,
-// and taking it down through nmcli would yank it out from under the tool that
-// owns it. Profiles the user actually imported are stored under `/etc`.
-function parseNmcliConnections(raw) {
-  var connections = []
-  var lines = String(raw || "").split("\n")
-  for (var i = 0; i < lines.length; i++) {
-    var line = lines[i]
-    if (line.trim() === "") continue
-
-    var fields = []
-    var rest = line
-    for (var f = 0; f < 4; f++) {
-      var pair = splitNmcliLine(rest)
-      fields.push(pair[0])
-      rest = pair[1]
-    }
-    fields.push(unescapeNmcli(rest))
-
-    if (fields[2] !== "vpn" && fields[2] !== "wireguard") continue
-    if (isVolatileConnection(fields[4])) continue
-    connections.push({
-      name: fields[0],
-      uuid: fields[1],
-      // "vpn" here means "needs the second pass to say which plugin".
-      kind: fields[2] === "wireguard" ? "wireguard" : "vpn",
-      active: fields[3] === "yes"
-    })
-  }
-  return connections
-}
-
-// `nmcli -t -f connection.uuid,vpn.service-type,vpn.data connection show <uuid>…`
-// prints one blank-line-separated block per connection, each line prefixed
-// with its field name. Returns { uuid: { serviceType, hasUsername } }.
-function parseNmcliVpnDetails(raw) {
-  var details = {}
-  var current = null
-  var lines = String(raw || "").split("\n")
-
-  function flush() {
-    if (current && current.uuid !== "") details[current.uuid] = current
-    current = null
-  }
-
-  for (var i = 0; i < lines.length; i++) {
-    var line = lines[i].trim()
-    if (line === "") {
-      flush()
-      continue
-    }
-
-    var pair = splitNmcliLine(line)
-    if (!current) current = { uuid: "", serviceType: "", hasUsername: false }
-
-    if (pair[0] === "connection.uuid") current.uuid = pair[1]
-    else if (pair[0] === "vpn.service-type") current.serviceType = pair[1]
-    else if (pair[0] === "vpn.data") current.hasUsername = hasVpnUsername(pair[1])
-  }
-  flush()
-
-  return details
-}
-
-// vpn.data is a comma-separated "key = value" list. OpenVPN's username lives
-// there rather than in vpn.secrets, so `nmcli --ask` never prompts for it —
-// a profile missing it authenticates as the empty user and is rejected.
-function hasVpnUsername(data) {
-  var entries = String(data || "").split(",")
-  for (var i = 0; i < entries.length; i++) {
-    var entry = entries[i].trim()
-    if (entry.indexOf("username") !== 0) continue
-
-    var value = entry.substring(entry.indexOf("=") + 1).trim()
-    if (value !== "") return true
-  }
-  return false
-}
-
-function isOpenVpnService(serviceType) {
-  return String(serviceType || "").toLowerCase().indexOf("openvpn") !== -1
-}
-
-function isWireGuard(profile) {
-  return profile && profile.kind === "wireguard"
-}
-
-function nmKindLabel(profile) {
-  return isWireGuard(profile) ? "WireGuard" : "OpenVPN"
-}
-
-// The glyph carries the kind, since the rows otherwise look identical and the
-// two behave differently the moment credentials come up.
-function nmTargets(profiles) {
-  var targets = []
-  for (var i = 0; i < profiles.length; i++) {
-    var profile = profiles[i]
-    var wireguard = isWireGuard(profile)
-
-    targets.push({
-      key: "profile:" + profile.uuid,
-      label: profile.name,
-      detail: profile.active
-        ? "Connected"
-        // Only OpenVPN can be missing a username: WireGuard keeps its keys in
-        // the profile, so there is nothing for the user to have left out.
-        : (wireguard || profile.hasUsername ? nmKindLabel(profile) + " profile" : "No username set"),
-      glyph: wireguard ? GLYPH_SHIELD : GLYPH_LOCK,
-      args: ["connection", "up", "uuid", profile.uuid],
-      uuid: profile.uuid,
-      kind: profile.kind,
-      hasUsername: profile.hasUsername
-    })
-  }
-  return targets
-}
-
-function nmSummary(profiles) {
-  for (var i = 0; i < profiles.length; i++) {
-    if (profiles[i].active) return profiles[i].name
-  }
-  return profiles.length === 0 ? "No profiles" : "Not connected"
-}
-
-function nmDetails(profiles) {
-  var rows = []
-  for (var i = 0; i < profiles.length; i++) {
-    if (!profiles[i].active) continue
-    rows.push(detail("Profile", profiles[i].name))
-    rows.push(detail("Type", nmKindLabel(profiles[i])))
-  }
-  if (rows.length > 0) rows.push(detail("Managed by", "NetworkManager"))
-  return rows
-}
-
-function activeNmProfile(profiles) {
-  for (var i = 0; i < profiles.length; i++) {
-    if (profiles[i].active) return profiles[i]
-  }
-  return null
-}
-
-// ---------------------------------------------------------------- Mullvad
-
-// `mullvad status -j` prints one JSON object. `state` is the tunnel state; the
-// shape of `details` follows it — an object while connected, connecting or
-// disconnected, a bare string ("reconnect") while disconnecting. Anything
-// unparseable is treated as "no idea", not as disconnected.
-function parseMullvadStatus(raw) {
-  var result = {
-    state: "",
-    connected: false,
-    blocked: false,
-    relay: "",
-    country: "",
-    city: "",
-    ip: "",
-    endpoint: "",
-    protocol: "",
-    tunnelInterface: "",
-    features: [],
-    lockedDown: false,
-    statusText: ""
-  }
-
-  var text = String(raw || "").trim()
-  if (text === "") return result
-
-  var payload = null
-  try {
-    payload = JSON.parse(text)
-  } catch (error) {
-    return result
-  }
-  if (!payload || typeof payload !== "object") return result
-
-  result.state = String(payload.state || "")
-  result.connected = result.state === "connected"
-  result.blocked = result.state === "error"
-
-  var details = payload.details
-  if (details && typeof details === "object") {
-    var location = details.location
-    if (location && typeof location === "object") {
-      result.relay = String(location.hostname || "")
-      result.country = String(location.country || "")
-      result.city = String(location.city || "")
-      result.ip = String(location.ipv4 || location.ipv6 || "")
-    }
-
-    var endpoint = details.endpoint
-    if (endpoint && typeof endpoint === "object") {
-      result.endpoint = String(endpoint.address || "")
-      result.protocol = String(endpoint.protocol || "").toUpperCase()
-      result.tunnelInterface = String(endpoint.tunnel_interface || "")
-    }
-
-    // Only reported while down, and only then does it mean traffic is being
-    // dropped right now. The lockdown *setting* is read separately.
-    if (details.locked_down === true) result.lockedDown = true
-
-    var indicators = details.feature_indicators
-    if (indicators && indicators.length) {
-      for (var i = 0; i < indicators.length; i++) result.features.push(mullvadFeatureLabel(indicators[i]))
-    }
-  }
-
-  result.statusText = mullvadStateText(result)
-  return result
-}
-
-// Feature indicators come back as CamelCase tags: "QuantumResistance".
-function mullvadFeatureLabel(tag) {
-  return String(tag || "").replace(/([a-z0-9])([A-Z])/g, "$1 $2")
-}
-
-function mullvadStateText(status) {
-  if (status.state === "connected") return "Connected"
-  if (status.state === "connecting") return "Connecting"
-  if (status.state === "disconnecting") return "Disconnecting"
-  if (status.state === "error") return "Blocked"
-  if (status.state === "disconnected") return "Disconnected"
-  return "Unknown"
-}
-
-function mullvadLocation(status) {
-  var parts = []
-  if (status.city !== "") parts.push(status.city)
-  if (status.country !== "") parts.push(status.country)
-  return parts.join(", ")
-}
-
-function mullvadSummary(status) {
-  var location = mullvadLocation(status)
-
-  if (status.state === "connected") {
-    if (location !== "" && status.relay !== "") return status.relay + " · " + location
-    if (location !== "") return location
-    if (status.relay !== "") return status.relay
-    return "Connected"
-  }
-  if (status.state === "connecting") return location !== "" ? "Connecting to " + location + "…" : "Connecting…"
-  if (status.state === "disconnecting") return "Disconnecting…"
-  // Not a quieter kind of disconnected: nothing is leaving the machine at all.
-  if (status.state === "error") return "Blocked — no traffic is leaving this machine"
-  if (status.state === "disconnected") return status.lockedDown ? "Not connected · traffic blocked" : "Not connected"
-  // No state yet: the first status call has not come back.
-  return "Checking…"
-}
-
-// Lockdown gets no row of its own: the daemon already reports it as a feature
-// indicator whenever it is on.
-function mullvadDetails(status) {
-  if (status.state !== "connected") return []
-
-  var rows = [detail("Relay", status.relay), detail("Location", mullvadLocation(status))]
-  if (status.endpoint !== "") {
-    rows.push(detail("Endpoint", status.protocol !== "" ? status.endpoint + " · " + status.protocol : status.endpoint))
-  }
-  if (status.tunnelInterface !== "") rows.push(detail("Interface", status.tunnelInterface))
-  if (status.features.length > 0) rows.push(detail("Features", status.features.join(", ")))
-  return rows.filter(function(row) { return row.value !== "" })
-}
-
-// Counts a tab as one level and four spaces as the same, since the CLI indents
-// with tabs but nothing promises it always will.
-function mullvadIndentDepth(line) {
-  var tabs = 0
-  var spaces = 0
-  for (var i = 0; i < line.length; i++) {
-    if (line[i] === "\t") { tabs += 1; continue }
-    if (line[i] === " ") { spaces += 1; continue }
-    break
-  }
-  return tabs + Math.floor(spaces / 4)
-}
-
-// `mullvad relay list` nests three levels, and every level looks like
-// "<name> (<code>)", so depth is what tells them apart:
-//
-//   Switzerland (ch)
-//   \tZurich (zrh) @ 47.36667°N, 8.55000°W
-//   \t\tch-zrh-wg-001 (185.156.46.146, …) - hosted by 31173 (rented)
-//
-// Only countries and cities are kept; individual relays are more choice than
-// a bar popup wants to offer.
-function parseMullvadRelays(raw) {
-  var countries = []
-  var current = null
-  var lines = String(raw || "").split("\n")
-
-  for (var i = 0; i < lines.length; i++) {
-    var line = lines[i]
-    if (line.trim() === "") continue
-
-    var depth = mullvadIndentDepth(line)
-    if (depth > 1) continue
-
-    var match = line.trim().match(/^(.+?)\s+\(([A-Za-z]{2,3})\)/)
-    if (!match) continue
-
-    var name = match[1].trim()
-    var code = match[2].toLowerCase()
-
-    if (depth === 0) {
-      if (code.length !== 2) continue
-      current = { name: name, code: code.toUpperCase(), cities: [] }
-      countries.push(current)
-    } else if (current && code.length === 3) {
-      current.cities.push({ name: name, code: code })
-    }
-  }
-  return countries
-}
-
-// Mullvad has no "fastest server" verb — `connect` uses whatever constraint is
-// stored — so the quick row is the widest constraint there is.
-function mullvadQuickTargets() {
-  return [
-    { key: "any", label: "Any location", detail: "Let Mullvad pick the relay", glyph: GLYPH_BOLT, args: ["any"] }
-  ]
-}
-
-// Favorites first, then the rest, filtering over the country name, its code,
-// and its city names so "zurich" finds Switzerland.
-function mullvadCountryTargets(countries, favorites, filter) {
-  var needle = String(filter || "").trim().toLowerCase()
-  var favored = []
-  var rest = []
-
-  for (var i = 0; i < countries.length; i++) {
-    var country = countries[i]
-    var cities = country.cities || []
-
-    if (needle !== "") {
-      var haystack = (country.name + " " + country.code + " " + cities.map(function(city) {
-        return city.name + " " + city.code
-      }).join(" ")).toLowerCase()
-      if (haystack.indexOf(needle) < 0) continue
-    }
-
-    var target = {
-      key: "country:" + country.code,
-      label: country.name,
-      detail: cities.length > 0 ? country.code + " · " + cities.length + (cities.length === 1 ? " city" : " cities") : country.code,
-      glyph: GLYPH_VPN,
-      args: [country.code.toLowerCase()]
-    }
-    if (needle === "" && favorites.indexOf(country.code) >= 0) favored.push(target)
-    else rest.push(target)
-  }
-
-  return favored.concat(rest)
-}
-
-// Which row to tick. The relay hostname carries its country ("ch-zrh-wg-504"),
-// which is exact; falling back to the country name covers the moment during a
-// connect when the hostname is not reported yet.
-function mullvadCurrentKey(status, countries) {
-  if (status.state !== "connected" && status.state !== "connecting") return ""
-
-  var hosted = String(status.relay || "").match(/^([a-z]{2})-/i)
-  if (hosted) return "country:" + hosted[1].toUpperCase()
-
-  var name = String(status.country || "").trim().toLowerCase()
-  if (name === "") return ""
-  for (var i = 0; i < countries.length; i++) {
-    if (countries[i].name.toLowerCase() === name) return "country:" + countries[i].code
-  }
-  return ""
-}
-
-// The three switchable settings live behind three separate subcommands, so the
-// backend asks for them in one shell and this reads whichever lines came back.
-// Each is matched on its leading words rather than on line order:
-//
-//   Autoconnect: off
-//   Block traffic when the VPN is disconnected: off
-//   Local network sharing setting: block
-//
-// One shell means one exit code — the last subcommand's — so a subcommand that
-// failed or that this version of the CLI does not have leaves no line and no
-// error. `seen` is which answers actually arrived: an unanswered setting is
-// unknown, not off. Reporting lockdown mode as off while it is on is the one
-// mistake a VPN widget must not make.
-function parseMullvadSettings(raw) {
-  var settings = { autoconnect: false, lockdown: false, lan: false, seen: {}, loaded: false }
-  var lines = String(raw || "").split("\n")
-
-  for (var i = 0; i < lines.length; i++) {
-    var line = lines[i].trim()
-    var separator = line.indexOf(":")
-    if (separator < 0) continue
-
-    var key = line.substring(0, separator).trim().toLowerCase()
-    var value = line.substring(separator + 1).trim().toLowerCase()
-    if (value === "") continue
-
-    if (key === "autoconnect") {
-      settings.autoconnect = value === "on"
-      settings.seen.autoconnect = true
-    } else if (key.indexOf("block traffic") === 0) {
-      settings.lockdown = value === "on"
-      settings.seen.lockdown = true
-    } else if (key.indexOf("local network sharing") === 0) {
-      settings.lan = value === "allow"
-      settings.seen.lan = true
-    }
-  }
-
-  settings.loaded = settings.seen.autoconnect === true
-    || settings.seen.lockdown === true
-    || settings.seen.lan === true
-  return settings
-}
-
-// A switch is drawn only for a setting the daemon answered for. A partial read
-// shows fewer switches rather than a full row of confident wrong ones.
-function mullvadToggles(settings) {
-  if (!settings || !settings.loaded) return []
-
-  var seen = settings.seen || {}
-  var toggles = []
-  if (seen.autoconnect === true) {
-    toggles.push(toggle("autoconnect", "Connect on startup", "Up as soon as the daemon starts", settings.autoconnect))
-  }
-  if (seen.lockdown === true) {
-    toggles.push(toggle("lockdown", "Lockdown mode", "No traffic at all while down", settings.lockdown))
-  }
-  if (seen.lan === true) {
-    toggles.push(toggle("lan", "Allow local network", "Reach printers and NAS while up", settings.lan))
-  }
-  return toggles
-}
-
-function mullvadToggleArgs(key, value) {
-  if (key === "autoconnect") return ["auto-connect", "set", value ? "on" : "off"]
-  if (key === "lockdown") return ["lockdown-mode", "set", value ? "on" : "off"]
-  if (key === "lan") return ["lan", "set", value ? "allow" : "block"]
-  return []
-}
-
-// ------------------------------------------------------------- Windscribe
-
-// `windscribe-cli status` prints a plain-text block whose lines come and go
-// with the state — `Protocol` only appears while a tunnel is up:
-//
-//   Internet connectivity: available
-//   Login state: Logged in
-//   Firewall state: On
-//   Connect state: Connected: Zurich - Alphorn
-//   Protocol: WireGuard:443
-//   Data usage: 96.99 MB / 2.00 GB
-//   Update available: 2.23.11
-//
-// So this matches on the leading key rather than on line position. Two values
-// carry a colon of their own — `Connect state` and `Protocol` — which is why
-// the split is on the first one only.
-function parseWindscribeStatus(raw) {
-  var result = {
-    loaded: false,
-    loggedIn: false,
-    loginText: "",
-    online: true,
-    firewallKnown: false,
-    firewallOn: false,
-    firewallAlways: false,
-    state: "",
-    connected: false,
-    location: "",
-    city: "",
-    nickname: "",
-    protocol: "",
-    port: "",
-    dataUsed: "",
-    dataLimit: "",
-    interference: false,
-    error: "",
-    statusText: ""
-  }
-
-  var lines = String(raw || "").split("\n")
-  for (var i = 0; i < lines.length; i++) {
-    var line = lines[i].trim()
-    if (line === "") continue
-
-    var separator = line.indexOf(":")
-    if (separator < 0) continue
-
-    var key = line.substring(0, separator).trim().toLowerCase()
-    var value = line.substring(separator + 1).trim()
-    if (value === "") continue
-
-    if (key === "login state") {
-      result.loginText = value
-      result.loggedIn = /^logged in/i.test(value)
-      result.loaded = true
-    } else if (key === "internet connectivity") {
-      result.online = !/^no|^unavailable|^not available/i.test(value)
-    } else if (key === "firewall state") {
-      // Three spellings, and only the two known ones are believed: reporting a
-      // kill switch as off while it is on is the one mistake worth ruling out.
-      var firewall = value.toLowerCase()
-      if (firewall === "off") {
-        result.firewallKnown = true
-      } else if (firewall === "on" || firewall === "always on") {
-        result.firewallKnown = true
-        result.firewallOn = true
-        result.firewallAlways = firewall === "always on"
-      }
-    } else if (key === "connect state") {
-      applyWindscribeConnectState(result, value)
-      result.loaded = true
-    } else if (key === "protocol") {
-      var protocol = value.split(":")
-      result.protocol = protocol[0].trim()
-      result.port = protocol.length > 1 ? protocol[1].trim() : ""
-    } else if (key === "data usage") {
-      var usage = value.split("/")
-      result.dataUsed = usage[0].trim()
-      result.dataLimit = usage.length > 1 ? usage[1].trim() : ""
-    }
-  }
-
-  result.statusText = windscribeStateText(result)
-  return result
-}
-
-// The value of `Connect state` is a sentence rather than an enum: "Connected:
-// Zurich - Alphorn", "Connecting", "Error: Location does not exist or is
-// disabled", or "Disconnected due to reaching WireGuard key limit. …".
-//
-// The error branch is not decoration. `windscribe-cli connect -n` exits 0 the
-// moment the daemon accepts the request, so a connect that fails seconds later
-// — a free account reaching for a Pro location, say — is reported here and
-// nowhere else. The state is sticky: it stays until something succeeds.
-function applyWindscribeConnectState(result, value) {
-  var text = String(value || "").trim()
-
-  // Appended to the connected state when the daemon sees the tunnel struggling.
-  var interference = text.match(/^(.*?)\s*\[Network interference\]$/i)
-  if (interference) {
-    result.interference = true
-    text = interference[1].trim()
-  }
-
-  var failure = text.match(/^error\s*:\s*(.+)$/i)
-  if (failure) {
-    result.state = "error"
-    result.error = failure[1].trim()
-    return
-  }
-
-  if (/^connecting\b/i.test(text)) {
-    result.state = "connecting"
-    applyWindscribeLocation(result, afterFirstColon(text))
-    return
-  }
-  if (/^connected\b/i.test(text)) {
-    result.state = "connected"
-    result.connected = true
-    applyWindscribeLocation(result, afterFirstColon(text))
-    return
-  }
-  if (/^disconnecting\b/i.test(text)) {
-    result.state = "disconnecting"
-    return
-  }
-  if (/^disconnected\b/i.test(text)) {
-    result.state = "disconnected"
-    // "Disconnected due to reaching WireGuard key limit. Use …" is a disconnect
-    // with a reason attached, and the reason is the half worth reading.
-    if (!/^disconnected$/i.test(text)) result.error = text
-  }
-}
-
-function afterFirstColon(text) {
-  var separator = String(text || "").indexOf(":")
-  return separator < 0 ? "" : text.substring(separator + 1).trim()
-}
-
-// `status` names the connected location as "City - Nickname", one segment
-// shorter than the three-part form `locations` prints, so the nickname is taken
-// from the end rather than from a fixed position.
-function applyWindscribeLocation(result, text) {
-  var location = String(text || "").trim()
-  if (location === "") return
-
-  result.location = location
-  var parts = location.split(" - ")
-  result.nickname = parts[parts.length - 1].trim()
-  if (parts.length > 1) result.city = parts[parts.length - 2].trim()
-}
-
-function windscribeStateText(status) {
-  if (!status.loaded) return "Unknown"
-  if (!status.loggedIn) return "Not logged in"
-  if (status.state === "connected") return "Connected"
-  if (status.state === "connecting") return "Connecting"
-  if (status.state === "disconnecting") return "Disconnecting"
-  if (status.state === "error") return "Not connected"
-  if (status.state === "disconnected") return "Disconnected"
-  return "Unknown"
-}
-
-function windscribeLocation(status) {
-  if (status.nickname !== "" && status.city !== "") return status.nickname + " · " + status.city
-  if (status.nickname !== "") return status.nickname
-  return status.city
-}
-
-function windscribeSummary(status) {
-  if (!status.loaded) return "Checking…"
-  if (!status.loggedIn) return "Not logged in"
-
-  if (status.state === "connected") {
-    var place = windscribeLocation(status)
-    var line = place !== "" ? place : "Connected"
-    return status.interference ? line + " · network interference" : line
-  }
-  if (status.state === "connecting") {
-    var target = windscribeLocation(status)
-    return target !== "" ? "Connecting to " + target + "…" : "Connecting…"
-  }
-  if (status.state === "disconnecting") return "Disconnecting…"
-  // The failure itself goes to lastError, which is where the panel puts the
-  // things a user has to act on; the summary only says where that leaves them.
-  if (status.state === "error") return status.firewallOn ? "Not connected · traffic blocked" : "Not connected"
-  if (status.state === "disconnected") return status.firewallOn ? "Not connected · traffic blocked" : "Not connected"
-  return "Checking…"
-}
-
-function windscribeDetails(status) {
-  if (status.state !== "connected") return []
-
-  var rows = [detail("Server", status.nickname), detail("Location", status.city)]
-  if (status.protocol !== "") {
-    rows.push(detail("Protocol", status.port !== "" ? status.protocol + " · " + status.port : status.protocol))
-  }
-  // Windscribe meters the free plan, and the number is only ever printed here.
-  if (status.dataUsed !== "") {
-    rows.push(detail("Data used", status.dataLimit !== "" ? status.dataUsed + " of " + status.dataLimit : status.dataUsed))
-  }
-  if (status.interference) rows.push(detail("Network", "Interference detected"))
-  return rows.filter(function(row) { return row.value !== "" })
-}
-
-// `windscribe-cli locations` prints one location per line:
-//
-//   Best Location - Alphorn (10 Gbps)
-//   US Central - Atlanta - Magic City (Pro) (10 Gbps)
-//   Switzerland - Zurich - Alphorn (10 Gbps)
-//
-// Region, city and nickname are joined with " - ", and the trailing groups
-// describe the entry rather than name it. Only the markers the CLI actually
-// prints are stripped, so a nickname that happens to end in brackets keeps
-// them.
-function parseWindscribeLocations(raw) {
-  var locations = []
-  var lines = String(raw || "").split("\n")
-
-  for (var i = 0; i < lines.length; i++) {
-    var line = lines[i].trim()
-    // `locations static` prefixes a device-name line and both variants say so
-    // when there is nothing to list.
-    if (line === "" || line === "No locations." || line.indexOf("(Device name:") === 0) continue
-
-    var pro = false
-    var disabled = false
-    var text = line
-
-    for (;;) {
-      var marker = text.match(/\s*\((Pro|Disabled|[^()]*Gbps)\)$/i)
-      if (marker) {
-        var name = marker[1].toLowerCase()
-        if (name === "pro") pro = true
-        else if (name === "disabled") disabled = true
-        text = text.substring(0, text.length - marker[0].length)
-        continue
-      }
-      // The CLI puts a static IP's device name in brackets.
-      var bracket = text.match(/\s*\[[^\[\]]*\]$/)
-      if (!bracket) break
-      text = text.substring(0, text.length - bracket[0].length)
-    }
-
-    var parts = text.split(" - ")
-    for (var p = 0; p < parts.length; p++) parts[p] = parts[p].trim()
-
-    var nickname = parts[parts.length - 1]
-    if (nickname === "") continue
-
-    locations.push({
-      region: parts.length > 1 ? parts[0] : "",
-      city: parts.length > 2 ? parts.slice(1, parts.length - 1).join(" - ") : "",
-      nickname: nickname,
-      pro: pro,
-      disabled: disabled
-    })
-  }
-  return locations
-}
-
-// The first row is the pseudo-location the daemon resolves "best" to. It names
-// a server rather than a place, so it belongs on the quick row and not in the
-// region list.
-function isWindscribeBestRow(location) {
-  return String(location.region || "").toLowerCase() === "best location"
-}
-
-function windscribeBestNickname(locations) {
-  for (var i = 0; i < locations.length; i++) {
-    if (isWindscribeBestRow(locations[i])) return locations[i].nickname
-  }
-  return ""
-}
-
-// Windscribe's own top-level grouping — a country most of the time, a slice of
-// one ("US East") where a country has too many. `connect` takes a region name
-// directly, so the row needs nothing beyond its name to be connectable, and
-// the daemon picks a datacenter inside it.
-function windscribeRegions(locations) {
-  var regions = []
-  var index = {}
-
-  for (var i = 0; i < locations.length; i++) {
-    var location = locations[i]
-    if (isWindscribeBestRow(location) || location.disabled || location.region === "") continue
-    // The name is handed to `windscribe-cli connect` as an argument. No shell is
-    // involved, so there is nothing to inject, but a name starting with a dash
-    // would be read as an option rather than as a place. No such region exists;
-    // a row that claimed to be one would not be a region either.
-    if (location.region.charAt(0) === "-") continue
-
-    var key = location.region.toLowerCase()
-    var region = index[key]
-    if (!region) {
-      region = { name: location.region, cities: [], nicknames: [], free: 0, total: 0 }
-      index[key] = region
-      regions.push(region)
-    }
-
-    if (location.city !== "" && region.cities.indexOf(location.city) < 0) region.cities.push(location.city)
-    region.nicknames.push(location.nickname)
-    region.total += 1
-    if (!location.pro) region.free += 1
-  }
-  return regions
-}
-
-// A free account cannot reach a Pro location, and the CLI's refusal — "Location
-// does not exist or is disabled" — does not say which of the two it meant. Say
-// it on the row instead of letting the connect fail to explain itself.
-function windscribeRegionDetail(region) {
-  var parts = []
-  if (region.cities.length > 0) {
-    parts.push(region.cities.length + (region.cities.length === 1 ? " city" : " cities"))
-  }
-  if (region.free === 0) parts.push("Pro only")
-  return parts.join(" · ")
-}
-
-// Windscribe names its regions and never prints a country code, so the
-// favorites setting is matched by name — "Switzerland" — and by leading word,
-// which is what makes "US" pick up US East, US Central and US West.
-function isWindscribeFavorite(name, favorites) {
-  if (!favorites || favorites.length === 0) return false
-
-  var region = String(name || "").toUpperCase()
-  var head = region.split(" ")[0]
-  for (var i = 0; i < favorites.length; i++) {
-    if (favorites[i] === region || favorites[i] === head) return true
-  }
-  return false
-}
-
-// `connect` with no location reuses the last one, which is the same row the
-// user just clicked; "best" is the only quick answer worth its own row.
-function windscribeQuickTargets(bestNickname) {
-  var nickname = String(bestNickname || "")
-  return [
-    {
-      key: "best",
-      label: "Best location",
-      detail: nickname !== "" ? "Currently " + nickname : "Let Windscribe pick",
-      glyph: GLYPH_BOLT,
-      args: ["best"]
-    }
-  ]
-}
-
-function windscribeRegionTargets(regions, favorites, filter) {
-  var needle = String(filter || "").trim().toLowerCase()
-  var favored = []
-  var rest = []
-
-  for (var i = 0; i < regions.length; i++) {
-    var region = regions[i]
-    if (needle !== "") {
-      // Nicknames are in the haystack because they are what the CLI reports
-      // back while connected, so "alphorn" finds the row it will tick.
-      var haystack = (region.name + " " + region.cities.join(" ") + " "
-        + region.nicknames.join(" ")).toLowerCase()
-      if (haystack.indexOf(needle) < 0) continue
-    }
-
-    var target = {
-      key: "region:" + region.name.toLowerCase(),
-      label: region.name,
-      detail: windscribeRegionDetail(region),
-      glyph: GLYPH_VPN,
-      args: [region.name]
-    }
-    if (needle === "" && isWindscribeFavorite(region.name, favorites)) favored.push(target)
-    else rest.push(target)
-  }
-
-  return favored.concat(rest)
-}
-
-// The status line names a city and a nickname but never the region, so the row
-// to tick is found by looking the connected server back up in the list. City
-// first, across every region, because a nickname is a joke and a city is not:
-// two regions sharing a nickname is likelier than two sharing a city.
-function windscribeCurrentKey(status, regions) {
-  if (status.state !== "connected" && status.state !== "connecting") return ""
-
-  var city = String(status.city || "").toLowerCase()
-  var nickname = String(status.nickname || "").toLowerCase()
-  var i = 0
-  var n = 0
-
-  if (city !== "") {
-    for (i = 0; i < regions.length; i++) {
-      for (n = 0; n < regions[i].cities.length; n++) {
-        if (regions[i].cities[n].toLowerCase() === city) return "region:" + regions[i].name.toLowerCase()
-      }
-    }
-  }
-  if (nickname !== "") {
-    for (i = 0; i < regions.length; i++) {
-      for (n = 0; n < regions[i].nicknames.length; n++) {
-        if (regions[i].nicknames[n].toLowerCase() === nickname) return "region:" + regions[i].name.toLowerCase()
-      }
-    }
-  }
-  return ""
-}
-
-// One switchable setting: the firewall, which is Windscribe's kill switch.
-// Everything else lives in the desktop app's preferences file, which this
-// widget reads nothing from and writes nothing to.
-function windscribeToggles(status) {
-  if (!status || !status.firewallKnown) return []
-
-  return [
-    toggle("firewall", "Firewall",
-      status.firewallAlways
-        ? "Always on — change it in the Windscribe app"
-        : "Block traffic outside the tunnel",
-      status.firewallOn)
-  ]
-}
-
-function windscribeToggleArgs(key, value) {
-  if (key === "firewall") return ["firewall", value ? "on" : "off"]
-  return []
-}
-
-// `windscribe-cli` holds a lock for as long as it runs, and a second
-// invocation does not wait its turn: it prints this and exits without doing
-// the work. The backend serialises its own calls, so seeing it means something
-// outside the widget is talking to the app — the user at a terminal, or this
-// same widget on another monitor, since each bar instantiates it separately.
-// It is not an answer about the tunnel, and nothing should be concluded from
-// it.
-function isWindscribeCliLocked(text) {
-  return /already running/i.test(String(text || ""))
-}
-
-// The firewall is a kill switch, and `lockdownMode` is read to decide whether
-// to warn someone that switching tools will not get through. "The tool never
-// said" has to weigh the same as "the tool said yes": the cost of a warning
-// nobody needed is a sentence, and the cost of the missing one is a connect
-// that fails for reasons the panel had already been told about and forgot.
-//
-// The summary line deliberately does not use this — it states what the tool
-// reported, and an unrecognised firewall line is not a report that traffic is
-// blocked. One weighs a risk, the other states a fact.
-function windscribeBlocksWhileDown(status) {
-  if (!status || !status.loaded) return false
-  return status.firewallOn || !status.firewallKnown
-}
-
-// --------------------------------------------------- Windscribe call queue
-//
-// Because two `windscribe-cli` invocations may not overlap, every call the
-// backend makes is a job in a queue of one runner. The rules are here rather
-// than in the QML so they can be tested: this is a reducer over (queue,
-// running job), not process plumbing, and it is the part that has been wrong.
-
-var WINDSCRIBE_MAX_ATTEMPTS = 4
-
-// Reads answer a question; actions change something. That is the whole
-// difference the queue cares about.
-function isWindscribeReadJob(kind) {
-  return kind === "status" || kind === "locations"
-}
-
-// Reads are idempotent, so a second one waiting behind the first is a stale
-// answer nobody asked for: the poll fires faster than the CLI replies, and a
-// backlog of them would keep the panel a cycle behind for as long as the shell
-// ran. A click goes to the front, because waiting out a two-hundred line
-// location fetch is not what "connect" should feel like.
-//
-// Returns the queue unchanged when the job is a duplicate read.
-function windscribeEnqueue(queue, runningKind, kind, args) {
-  var next = queue.slice()
-  var job = { kind: kind, args: args, attempts: 0 }
-
-  if (!isWindscribeReadJob(kind)) {
-    next.unshift(job)
-    return next
-  }
-
-  if (runningKind === kind) return queue
-  for (var i = 0; i < next.length; i++) {
-    if (next[i].kind === kind) return queue
-  }
-  next.push(job)
-  return next
-}
-
-// A job that lost the lock goes back to the head of its own queue, one attempt
-// older. A fresh object rather than a mutated one, so nothing that captured the
-// old job sees its attempt count move underneath it.
-function windscribeRequeue(queue, job) {
-  var next = queue.slice()
-  next.unshift({ kind: job.kind, args: job.args, attempts: job.attempts + 1 })
-  return next
-}
-
-// Bounded, because a lock held by something that is never going to let go is
-// not worth knocking on forever. Past the last attempt the job runs its
-// finisher, which treats a refusal as inconclusive rather than as news.
-function windscribeCanRetry(job) {
-  return !!job && job.attempts < WINDSCRIBE_MAX_ATTEMPTS
-}
-
-// The jitter is not decoration. Two widget instances that started together lose
-// together — that is what a shared lock and a shared poll interval do — and
-// backing off by the same amount would have them collide again on every round.
-// Something has to break the tie, and the attempt count cannot: it is identical
-// on both sides. `roll` is a 0..1 draw, passed in so this stays testable.
-function windscribeRetryDelay(attempts, roll) {
-  var draw = Number(roll)
-  if (!isFinite(draw) || draw < 0) draw = 0
-  if (draw > 1) draw = 1
-  return 120 * attempts + Math.floor(draw * 280)
-}
-
-// Whether a job of any of these kinds is running or waiting — what stops a
-// second click landing on top of the first.
-function windscribePending(queue, runningKind, kinds) {
-  if (kinds.indexOf(runningKind) >= 0) return true
-  for (var i = 0; i < queue.length; i++) {
-    if (kinds.indexOf(queue[i].kind) >= 0) return true
-  }
-  return false
-}
